Noted change of canonical repo
I recently changed my GitHub handle to @redoPop and created this
account and fork to make sure I don't break any prior references
to my code. The canonical version of this repository is now:

http://github.com/redoPop/loupe

Please use this URL (above) and not the repo you're currently
looking at. This version will not be maintained.